% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiverseworkshop4-package.R
\docType{package}
\name{epiverseworkshop4-package}
\alias{epiverseworkshop4}
\alias{epiverseworkshop4-package}
\title{epiverseworkshop4: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/scarabel/epiverseworkshop4}
  \item Report bugs at \url{https://github.com/scarabel/epiverseworkshop4/issues}
}

}
\author{
\strong{Maintainer}: Francesca Scarabel \email{francesca.scarabel@gmail.com} (\href{https://orcid.org/0000-0003-0250-4555}{ORCID}) [copyright holder]

}
\keyword{internal}
